+++ app/Models/CategoryRoom.php
@@ -3,6 +3,7 @@
 namespace App\Models;
 
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\Relations\HasMany;
 
 class CategoryRoom extends Model
 {
@@ -14,4 +15,9 @@ class CategoryRoom extends Model
         'created_at',
         'updated_at'
     ];
+
+    public function chatRoom(): HasMany
+    {
+        return $this->hasMany(chatRoom::class);
+    }
 }