+++ app/Models/ChatRoom.php
@@ -5,6 +5,7 @@ namespace App\Models;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Illuminate\Database\Eloquent\Relations\BelongsToMany;
+use Illuminate\Database\Eloquent\Relations\HasMany;
 
 class ChatRoom extends Model
 {
@@ -29,5 +30,11 @@ class ChatRoom extends Model
     public function usersList(): BelongsToMany
     {
         return $this->belongsToMany(User::class, 'users_has_chat_rooms', 'chat_rooms_id', 'users_id');
+    }
+
+    // Relacionamento One-to-Many com mensagens
+    public function messages(): HasMany
+    {
+        return $this->hasMany(Message::class);
     }
 }