+++ database/migrations/2025_06_25_134645_create_messages_table.php
@@ -1,0 +1,32 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+return new class extends Migration
+{
+    /**
+     * Run the migrations.
+     */
+    public function up(): void
+    {
+        Schema::create('messages', function (Blueprint $table) {
+            $table->id();
+            $table->text('message', 500);
+            $table->unsignedBigInteger('user_id');
+            $table->unsignedBigInteger('chat_room_id');
+            $table->datetime('send_at')->nullable();
+            $table->foreign('user_id')->references('id')->on('users');
+            $table->foreign('chat_room_id')->references('id')->on('chat_rooms');
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     */
+    public function down(): void
+    {
+        Schema::dropIfExists('messages');
+    }
+};